requirements/pip.txt
# Base packages
pip==10.0.0
appdirs==1.4.3
virtualenv==15.2.0
docutils==0.14
Sphinx==1.7.4

# 0.3.0 breaks sidebar
# https://github.com/rtfd/sphinx_rtd_theme/issues/610
sphinx_rtd_theme==0.2.5b1

Pygments==2.2.0

# latest compatible version with our code
# https://github.com/rtfd/readthedocs.org/pull/2916#discussion_r172991757
mkdocs==0.15.0
Markdown==2.6.11

django==1.9.13
six==1.11.0
future==0.16.0
readthedocs-build<2.1

# django-tastypie 0.13.x and 0.14.0 are not compatible with our code
django-tastypie==0.13.0

django-guardian==1.4.9
django-extensions==2.0.6

# djangorestframework 3.7.x drops support for django 1.9.x
djangorestframework==3.6.4

django-vanilla-views==1.0.5
jsonfield==2.0.2

requests==2.18.4
slumber==0.7.1
lxml==4.2.1
defusedxml==0.5.0

# Basic tools
redis==2.10.6
celery==4.1.0
kombu==4.1.0

# django-allauth 0.33.0 dropped support for Django 1.9
# https://django-allauth.readthedocs.io/en/latest/release-notes.html#backwards-incompatible-changes
django-allauth==0.32.0

dnspython==1.15.0

# VCS
httplib2==0.11.3
GitPython==2.1.9

# Search
elasticsearch==5.5.2
pyquery==1.4.0

# Utils
django-gravatar2==1.4.2
pytz==2018.4
beautifulsoup4==4.6.0
Unipath==1.1
django-kombu==0.9.4
mimeparse==0.1.3
mock==2.0.0

# stripe 1.20.2 is the latest compatible with our code base (otherwise
# gold/tests/test_forms.py fails)
stripe==1.20.2

django-formtools==2.1

# docker is pinned to 3.1.3 because we found some strange behavior
# related to timeouts on EXEC with 3.2.1 that's not present with 3.1.3
# https://github.com/rtfd/readthedocs.org/issues/3999
docker==3.1.3

django-textclassifier==1.0
django-annoying==0.10.4
django-messages-extends==0.6.0
djangorestframework-jsonp==1.0.2
django-taggit==0.22.2
dj-pagination==2.3.2
django-autocomplete-light==3.4.1

# Docs
sphinxcontrib-httpdomain==1.6.1

# commonmark 0.5.5 is the latest version compatible with our docs, the
# newer ones make `tox -e docs` to fail
commonmark==0.5.5

recommonmark==0.4.0

# Version comparison stuff
packaging==17.1

# Commenting stuff
django-cors-middleware==1.3.1
nilsimsa==0.3.8
